+++ src/diagnostics.rs
@@ -1,7 +1,10 @@
-use crate::fun::{display::DisplayFn, Name};
+use TSPL::ParseError;
+
+use crate::fun::{display::DisplayFn, Name, Source};
 use std::{
   collections::BTreeMap,
   fmt::{Display, Formatter},
+  ops::Range,
 };
 
 pub const ERR_INDENT_SIZE: usize = 2;
@@ -28,16 +31,19 @@ pub struct DiagnosticsConfig {
 
 #[derive(Debug, Clone)]
 pub struct Diagnostic {
-  message: String,
-  severity: Severity,
+  pub message: String,
+  pub severity: Severity,
+  pub source: Source,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum DiagnosticOrigin {
+  /// An error when parsing source code.
+  Parsing,
   /// An error from the relationship between multiple top-level definitions.
   Book,
-  /// An error in a pattern-matching function definition rule.
-  Rule(Name),
+  /// An error in a function definition.
+  Function(Name),
   /// An error in a compiled inet.
   Inet(String),
   /// An error during readback of hvm-core run results.
@@ -68,27 +74,48 @@ impl Diagnostics {
     Self { err_counter: 0, diagnostics: Default::default(), config }
   }
 
-  pub fn add_book_error(&mut self, err: impl std::fmt::Display) {
+  pub fn add_parsing_error(&mut self, err: impl std::fmt::Display, source: Source) {
     self.err_counter += 1;
-    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Book);
+    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Parsing, source);
   }
 
-  pub fn add_rule_error(&mut self, err: impl std::fmt::Display, def_name: Name) {
+  pub fn add_book_error(&mut self, err: impl std::fmt::Display) {
     self.err_counter += 1;
-    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Rule(def_name.def_name_from_generated()));
+    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Book, Default::default());
+  }
+
+  pub fn add_function_error(&mut self, err: impl std::fmt::Display, name: Name, source: Source) {
+    self.err_counter += 1;
+    self.add_diagnostic(
+      err,
+      Severity::Error,
+      DiagnosticOrigin::Function(name.def_name_from_generated()),
+      source,
+    );
   }
 
   pub fn add_inet_error(&mut self, err: impl std::fmt::Display, def_name: String) {
     self.err_counter += 1;
-    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Inet(def_name));
+    self.add_diagnostic(err, Severity::Error, DiagnosticOrigin::Inet(def_name), Default::default());
   }
 
-  pub fn add_rule_warning(&mut self, warn: impl std::fmt::Display, warn_type: WarningType, def_name: Name) {
+  pub fn add_function_warning(
+    &mut self,
+    warn: impl std::fmt::Display,
+    warn_type: WarningType,
+    def_name: Name,
+    source: Source,
+  ) {
     let severity = self.config.warning_severity(warn_type);
     if severity == Severity::Error {
       self.err_counter += 1;
     }
-    self.add_diagnostic(warn, severity, DiagnosticOrigin::Rule(def_name.def_name_from_generated()));
+    self.add_diagnostic(
+      warn,
+      severity,
+      DiagnosticOrigin::Function(def_name.def_name_from_generated()),
+      source,
+    );
   }
 
   pub fn add_book_warning(&mut self, warn: impl std::fmt::Display, warn_type: WarningType) {
@@ -96,11 +123,17 @@ impl Diagnostics {
     if severity == Severity::Error {
       self.err_counter += 1;
     }
-    self.add_diagnostic(warn, severity, DiagnosticOrigin::Book);
+    self.add_diagnostic(warn, severity, DiagnosticOrigin::Book, Default::default());
   }
 
-  pub fn add_diagnostic(&mut self, msg: impl ToString, severity: Severity, orig: DiagnosticOrigin) {
-    let diag = Diagnostic { message: msg.to_string(), severity };
+  pub fn add_diagnostic(
+    &mut self,
+    msg: impl std::fmt::Display,
+    severity: Severity,
+    orig: DiagnosticOrigin,
+    source: Source,
+  ) {
+    let diag = Diagnostic { message: msg.to_string(), severity, source };
     self.diagnostics.entry(orig).or_default().push(diag)
   }
 
@@ -112,7 +145,7 @@ impl Diagnostics {
     match result {
       Ok(t) => Some(t),
       Err(e) => {
-        self.add_rule_error(e, def_name);
+        self.add_function_error(e, def_name, Default::default());
         None
       }
     }
@@ -159,47 +192,110 @@ impl Diagnostics {
   /// Returns a Display that prints the diagnostics with one of the given severities.
   pub fn display_with_severity(&self, severity: Severity) -> impl std::fmt::Display + '_ {
     DisplayFn(move |f| {
-      fn filter<'a>(
-        errs: impl IntoIterator<Item = &'a Diagnostic>,
-        severity: Severity,
-      ) -> impl Iterator<Item = &'a Diagnostic> {
-        errs.into_iter().filter(move |err| err.severity == severity)
-      }
+      // We want to print diagnostics information somewhat like this:
+      // ```
+      // In file A :
+      // In definition X :
+      //   {error}
+      // In definition Y :
+      //   {error}
+      //
+      // In file B :
+      // In compiled Inet Z :
+      //   {error}
+      //
+      // Other diagnostics:
+      // In {...}
+      // ```
+      // The problem is, diagnostics data is currently structured as a mapping from something like
+      // DiagnosticOrigin to Vec<(DiagnosticMessage, DiagnosticFile)>, and we would need something
+      // like a mapping from DiagnosticFile to DiagnosticOrigin to Vec<DiagnosticMessage> in order
+      // to print it cleanly. We might want to change it later to have this structure,
+      // but meanwhile, we do the transformations below to make the goal possible.
 
-      let mut has_msg = false;
-      for (orig, errs) in &self.diagnostics {
-        let mut errs = filter(errs, severity).peekable();
-        if errs.peek().is_some() {
-          match orig {
-            DiagnosticOrigin::Book => {
-              for err in errs {
-                writeln!(f, "{err}")?;
-              }
-            }
-            DiagnosticOrigin::Rule(nam) => {
-              writeln!(f, "\x1b[1mIn definition '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
-              for err in errs {
-                writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
-              }
-            }
-            DiagnosticOrigin::Inet(nam) => {
-              writeln!(f, "\x1b[1mIn compiled inet '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
-              for err in errs {
-                writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
-              }
-            }
-            DiagnosticOrigin::Readback => {
-              writeln!(f, "\x1b[1mDuring readback:\x1b[0m")?;
-              for err in errs {
-                writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
-              }
-            }
+      // Ignore diagnostics without the desired severity.
+      let diagnostics = self
+        .diagnostics
+        .iter()
+        .map(|(origin, diags)| (origin, diags.iter().filter(|diag| diag.severity == severity)));
+
+      // Produce the structure described above.
+      let groups: BTreeMap<&Option<String>, BTreeMap<&DiagnosticOrigin, Vec<&Diagnostic>>> = diagnostics
+        .fold(BTreeMap::new(), |mut file_tree, (origin, diags)| {
+          for diag in diags {
+            // We need to allow this Clippy warning due to `Name` in `DiagnosticOrigin::Function`.
+            // We know how it works, so it shouldn't be a problem.
+            #[allow(clippy::mutable_key_type)]
+            let file_group_entry = file_tree.entry(&diag.source.file).or_default();
+            let origin_group_entry = file_group_entry.entry(origin).or_default();
+            origin_group_entry.push(diag);
           }
-          has_msg = true;
+          file_tree
+        });
+      // Now, we have a mapping from DiagnosticFile to DiagnosticOrigin to Vec<DiagnosticMessage>.
+
+      // If the last file is `None`, it means we only have diagnostics with unknown source file.
+      // In this case, we won't print a special message for them.
+      let only_unknown_file_diagnostics = groups.keys().next_back() == Some(&&None);
+
+      // Reverse the group iterator so `None` files go last.
+      for (file, origin_to_diagnostics) in groups.iter().rev() {
+        if !only_unknown_file_diagnostics {
+          match &file {
+            Some(name) => writeln!(f, "\x1b[1mIn \x1b[4m{}\x1b[0m\x1b[1m :\x1b[0m", name)?,
+            None => writeln!(f, "\x1b[1mOther diagnostics:\x1b[0m")?,
+          };
+        }
+
+        let mut has_msg = false;
+        for (origin, diagnostics) in origin_to_diagnostics {
+          let mut diagnostics = diagnostics.iter().peekable();
+          if diagnostics.peek().is_some() {
+            match origin {
+              DiagnosticOrigin::Parsing => {
+                for err in diagnostics {
+                  writeln!(f, "{err}")?;
+                }
+              }
+              DiagnosticOrigin::Book => {
+                for err in diagnostics {
+                  writeln!(f, "{err}")?;
+                }
+              }
+              DiagnosticOrigin::Function(nam) => {
+                writeln!(f, "\x1b[1mIn definition '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
+                for err in diagnostics {
+                  writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
+                }
+              }
+              DiagnosticOrigin::Inet(nam) => {
+                writeln!(f, "\x1b[1mIn compiled inet '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
+                for err in diagnostics {
+                  writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
+                }
+              }
+              DiagnosticOrigin::Readback => {
+                writeln!(f, "\x1b[1mDuring readback:\x1b[0m")?;
+                for err in diagnostics {
+                  writeln!(f, "{:ERR_INDENT_SIZE$}{err}", "")?;
+                }
+              }
+            }
+            has_msg = true;
+          }
+        }
+        if has_msg {
+          writeln!(f)?;
         }
       }
-      if has_msg {
-        writeln!(f)?;
+      Ok(())
+    })
+  }
+
+  pub fn display_only_messages(&self) -> impl std::fmt::Display + '_ {
+    DisplayFn(move |f| {
+      for err in self.diagnostics.values().flatten() {
+        writeln!(f, "{err}")?;
       }
       Ok(())
     })
@@ -223,7 +319,24 @@ impl From<String> for Diagnostics {
     Self {
       diagnostics: BTreeMap::from_iter([(
         DiagnosticOrigin::Book,
-        vec![Diagnostic { message: value, severity: Severity::Error }],
+        vec![Diagnostic { message: value, severity: Severity::Error, source: Default::default() }],
+      )]),
+      ..Default::default()
+    }
+  }
+}
+
+impl From<ParseError> for Diagnostics {
+  /// Transforms a parse error into `Diagnostics`.
+  ///
+  /// NOTE: Since `ParseError` does not include the source code, we can't get the `TextLocation` of the error,
+  /// so it is not included in the diagnostic.
+  /// range is set as None.
+  fn from(value: ParseError) -> Self {
+    Self {
+      diagnostics: BTreeMap::from_iter([(
+        DiagnosticOrigin::Parsing,
+        vec![Diagnostic { message: value.into(), severity: Severity::Error, source: Default::default() }],
       )]),
       ..Default::default()
     }
@@ -271,5 +384,109 @@ impl Default for DiagnosticsConfig {
 impl Display for Diagnostic {
   fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
     write!(f, "{}", self.message)
+  }
+}
+
+impl Diagnostic {
+  pub fn display_with_origin<'a>(&'a self, origin: &'a DiagnosticOrigin) -> impl std::fmt::Display + '_ {
+    DisplayFn(move |f| {
+      match origin {
+        DiagnosticOrigin::Parsing => writeln!(f, "{self}")?,
+        DiagnosticOrigin::Book => writeln!(f, "{self}")?,
+        DiagnosticOrigin::Function(nam) => {
+          writeln!(f, "\x1b[1mIn definition '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
+          writeln!(f, "{:ERR_INDENT_SIZE$}{self}", "")?;
+        }
+        DiagnosticOrigin::Inet(nam) => {
+          writeln!(f, "\x1b[1mIn compiled inet '\x1b[4m{}\x1b[0m\x1b[1m':\x1b[0m", nam)?;
+          writeln!(f, "{:ERR_INDENT_SIZE$}{self}", "")?;
+        }
+        DiagnosticOrigin::Readback => {
+          writeln!(f, "\x1b[1mDuring readback:\x1b[0m")?;
+          writeln!(f, "{:ERR_INDENT_SIZE$}{self}", "")?;
+        }
+      };
+      Ok(())
+    })
+  }
+}
+
+#[derive(Debug, Copy, Clone, Hash, PartialEq, PartialOrd, Ord, Eq)]
+pub struct TextLocation {
+  pub line: usize,
+  pub char: usize,
+}
+
+impl TextLocation {
+  pub fn new(line: usize, char: usize) -> Self {
+    TextLocation { line, char }
+  }
+
+  /// Transforms a `usize` byte index on `code` into a `TextLocation`.
+  pub fn from_byte_loc(code: &str, loc: usize) -> Self {
+    let code = code.as_bytes();
+    let mut line = 0;
+    let mut char = 0;
+    let mut cur_idx = 0;
+    while cur_idx < loc && cur_idx < code.len() {
+      if code[cur_idx] == b'\n' {
+        line += 1;
+        char = 0;
+      } else {
+        char += 1;
+      }
+      cur_idx += 1;
+    }
+
+    TextLocation { line, char }
+  }
+}
+
+#[derive(Debug, Copy, Clone, Hash, PartialEq, PartialOrd, Ord, Eq)]
+pub struct TextSpan {
+  pub start: TextLocation,
+  pub end: TextLocation,
+}
+
+impl TextSpan {
+  pub fn new(start: TextLocation, end: TextLocation) -> Self {
+    TextSpan { start, end }
+  }
+
+  /// Transforms a `usize` byte range on `code` into a `TextLocation`.
+  pub fn from_byte_span(code: &str, span: Range<usize>) -> Self {
+    // Will loop for way too long otherwise
+    assert!(span.start <= span.end);
+
+    let code = code.as_bytes();
+    let mut start_line = 0;
+    let mut start_char = 0;
+    let mut end_line;
+    let mut end_char;
+
+    let mut cur_idx = 0;
+    while cur_idx < span.start && cur_idx < code.len() {
+      if code[cur_idx] == b'\n' {
+        start_line += 1;
+        start_char = 0;
+      } else {
+        start_char += 1;
+      }
+      cur_idx += 1;
+    }
+
+    end_line = start_line;
+    end_char = start_char;
+    while cur_idx < span.end && cur_idx < code.len() {
+      if code[cur_idx] == b'\n' {
+        end_line += 1;
+        end_char = 0;
+      } else {
+        end_char += 1;
+      }
+      cur_idx += 1;
+    }
+
+    TextSpan::new(TextLocation::new(start_line, start_char), TextLocation::new(end_line, end_char))
   }
 }