+++ src/imp/to_fun.rs
@@ -1,13 +1,17 @@
 use super::{AssignPattern, Definition, Expr, InPlaceOp, Stmt};
-use crate::fun::{
-  self,
-  builtins::{LCONS, LNIL},
-  parser::ParseBook,
-  Book, Name,
+use crate::{
+  diagnostics::Diagnostics,
+  fun::{
+    self,
+    builtins::{LCONS, LNIL},
+    parser::ParseBook,
+    Book, Name,
+  },
 };
 
 impl ParseBook {
-  pub fn to_fun(mut self) -> Result<Book, String> {
+  // TODO: Change all functions to return diagnostics
+  pub fn to_fun(mut self) -> Result<Book, Diagnostics> {
     for (name, mut def) in std::mem::take(&mut self.imp_defs) {
       def.order_kwargs(&self)?;
       def.gen_map_get();
@@ -25,12 +29,23 @@ impl ParseBook {
 }
 
 impl Definition {
-  pub fn to_fun(self) -> Result<fun::Definition, String> {
-    let body = self.body.into_fun().map_err(|e| format!("In function '{}': {}", self.name, e))?;
+  pub fn to_fun(self) -> Result<fun::Definition, Diagnostics> {
+    let body = self.body.into_fun().map_err(|e| {
+      let mut diags = Diagnostics::default();
+      diags.add_function_error(e, self.name.clone(), self.source.clone());
+      diags
+    })?;
+
     let body = match body {
       StmtToFun::Return(term) => term,
       StmtToFun::Assign(..) => {
-        return Err(format!("Function '{}' doesn't end with a return statement", self.name));
+        let mut diags = Diagnostics::default();
+        diags.add_function_error(
+          "Function doesn't end with a return statement",
+          self.name,
+          self.source.clone(),
+        );
+        return Err(diags);
       }
     };
 
@@ -100,7 +115,7 @@ impl Stmt {
         wrap(nxt_pat, term, ask)
       }
       Stmt::Assign { pat: AssignPattern::MapSet(..), val: _, nxt: None } => {
-        return Err("Branch ends with map assignment.".to_string());
+        return Err("Branch ends with map assignment.".to_string())?;
       }
       Stmt::Assign { pat, val, nxt: Some(nxt) } => {
         let pat = pat.into_fun();
@@ -179,17 +194,17 @@ impl Stmt {
             (ask && ask_, Some(tp), t, e)
           }
           (StmtToFun::Assign(..), StmtToFun::Assign(..)) => {
-            return Err("'if' branches end with different assignments.".to_string());
+            return Err("'if' branches end with different assignments.".to_string())?;
           }
           (StmtToFun::Return(..), StmtToFun::Assign(..)) => {
             return Err(
               "Expected 'else' branch from 'if' to return, but it ends with assignment.".to_string(),
-            );
+            )?;
           }
           (StmtToFun::Assign(..), StmtToFun::Return(..)) => {
             return Err(
               "Expected 'else' branch from 'if' to end with assignment, but it returns.".to_string(),
-            );
+            )?;
           }
         };
         let arms = vec![else_, then];
@@ -215,13 +230,13 @@ impl Stmt {
           let (arm_ask, arm_pat, arm_rgt) = take(arm.rgt)?;
           match (&arm_pat, &fst_pat) {
             (Some(arm_pat), Some(fst_pat)) if arm_pat != fst_pat || arm_ask != fst_ask => {
-              return Err("'match' arms end with different assignments.".to_string());
+              return Err("'match' arms end with different assignments.".to_string())?;
             }
             (Some(_), None) => {
-              return Err("Expected 'match' arms to end with assignment, but it returns.".to_string());
+              return Err("Expected 'match' arms to end with assignment, but it returns.".to_string())?;
             }
             (None, Some(_)) => {
-              return Err("Expected 'match' arms to return, but it ends with assignment.".to_string());
+              return Err("Expected 'match' arms to return, but it ends with assignment.".to_string())?;
             }
             (Some(_), Some(_)) => fun_arms.push((arm.lft, vec![], arm_rgt)),
             (None, None) => fun_arms.push((arm.lft, vec![], arm_rgt)),
@@ -242,13 +257,13 @@ impl Stmt {
           let (arm_ask, arm_pat, arm) = take(arm)?;
           match (&arm_pat, &fst_pat) {
             (Some(arm_pat), Some(fst_pat)) if arm_pat != fst_pat || arm_ask != fst_ask => {
-              return Err("'switch' arms end with different assignments.".to_string());
+              return Err("'switch' arms end with different assignments.".to_string())?;
             }
             (Some(_), None) => {
-              return Err("Expected 'switch' arms to end with assignment, but it returns.".to_string());
+              return Err("Expected 'switch' arms to end with assignment, but it returns.".to_string())?;
             }
             (None, Some(_)) => {
-              return Err("Expected 'switch' arms to return, but it ends with assignment.".to_string());
+              return Err("Expected 'switch' arms to return, but it ends with assignment.".to_string())?;
             }
             (Some(_), Some(_)) => fun_arms.push(arm),
             (None, None) => fun_arms.push(arm),
@@ -270,13 +285,13 @@ impl Stmt {
           let (arm_ask, arm_pat, arm_rgt) = take(arm.rgt)?;
           match (&arm_pat, &fst_pat) {
             (Some(arm_pat), Some(fst_pat)) if arm_pat != fst_pat || arm_ask != fst_ask => {
-              return Err("'fold' arms end with different assignments.".to_string());
+              return Err("'fold' arms end with different assignments.".to_string())?;
             }
             (Some(_), None) => {
-              return Err("Expected 'fold' arms to end with assignment, but it returns.".to_string());
+              return Err("Expected 'fold' arms to end with assignment, but it returns.".to_string())?;
             }
             (None, Some(_)) => {
-              return Err("Expected 'fold' arms to return, but it ends with assignment.".to_string());
+              return Err("Expected 'fold' arms to return, but it ends with assignment.".to_string())?;
             }
             (Some(_), Some(_)) => fun_arms.push((arm.lft, vec![], arm_rgt)),
             (None, None) => fun_arms.push((arm.lft, vec![], arm_rgt)),
@@ -294,17 +309,17 @@ impl Stmt {
             (aa && ba, Some(sp), s, b)
           }
           (StmtToFun::Assign(..), StmtToFun::Assign(..)) => {
-            return Err("'bend' branches end with different assignments.".to_string());
+            return Err("'bend' branches end with different assignments.".to_string())?;
           }
           (StmtToFun::Return(..), StmtToFun::Assign(..)) => {
             return Err(
               "Expected 'else' branch from 'bend' to return, but it ends with assignment.".to_string(),
-            );
+            )?;
           }
           (StmtToFun::Assign(..), StmtToFun::Return(..)) => {
             return Err(
               "Expected 'else' branch from 'bend' to end with assignment, but it returns.".to_string(),
-            );
+            )?;
           }
         };
         let term =
@@ -340,7 +355,7 @@ impl Stmt {
       Stmt::Return { term } => StmtToFun::Return(term.to_fun()),
       Stmt::LocalDef { def, nxt } => {
         let (ask, nxt_pat, nxt) = take(*nxt)?;
-        let def = def.to_fun()?;
+        let def = def.to_fun().map_err(|e| e.display_only_messages().to_string())?;
         let term = fun::Term::Def { def, nxt: Box::new(nxt) };
         wrap(nxt_pat, term, ask)
       }
@@ -466,7 +481,7 @@ fn wrap_nxt_assign_stmt(
       };
       Ok(wrap(nxt_pat, term, ask_nxt))
     } else {
-      Err("Statement ends with return but is not at end of function.".to_string())
+      Err("Statement ends with return but is not at end of function.".to_string())?
     }
   } else if let Some(pat) = pat {
     Ok(StmtToFun::Assign(ask, pat, term))